Remove redundant ForkChoiceStrategy interface.
