+++ protoarray/src/test/java/tech/pegasys/teku/protoarray/NoVotesTest.java
@@ -35,8 +35,7 @@ public class NoVotesTest {
   void noVotesTest() {
     VoteUpdater store = createStoreToManipulateVotes();
 
-    ProtoArrayForkChoiceStrategy forkChoice =
-        createProtoArrayForkChoiceStrategy(getHash(0), ZERO, ONE, ONE);
+    ForkChoiceStrategy forkChoice = createProtoArrayForkChoiceStrategy(getHash(0), ZERO, ONE, ONE);
 
     List<UInt64> balances = new ArrayList<>(Collections.nCopies(16, ZERO));
